api/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy             # for orm
mysqlclient            # mysql connector
python-dotenv          # to store environment variables in .env file, we can store in system, but python-dotenv is recommended
pydantic[email]        # for email-validation
passlib                # for password hashing
aiofiles               # for serving static files
alembic		#for migration
PyJWT     #jwt